js: Don't use paster to build.
It's unnecessary to setup the Pylons app context for building.
