+++ r2/Makefile
@@ -186,7 +186,7 @@ define JS_MODULE_TEMPLATE
 $(JS_MODULE_OUTPUTS_$(1)): $(JS_MODULE_DEPS_$(1)) r2/lib/js.py
 	# remove mangled output symlinks, similar to above.
 	rm -f $(JS_MODULE_OUTPUTS_$(1))
-	paster run standalone r2/lib/js.py -c "build_module('$(1)')"
+	$(PYTHON) r2/lib/js.py build_module "$(1)"
 	touch $$@
 endef
